server_django/requirements.txt
django
djangorestframework
coreapi
requests
django-plotly-dash
django-guardian
# django_auth_ldap
